requirements.txt
# ============================================
# 茶品六因子 AI 评分器 Pro - Python 依赖
# Python 3.10+
# ============================================

# ============================================
# 核心框架
# ============================================
streamlit==1.55.0              # Web 应用框架
requests==2.32.5               # HTTP 请求库

# ============================================
# AI & LLM
# ============================================
openai==2.26.0                 # OpenAI API 客户端
dashscope==1.25.13             # 阿里云 Dashscope 服务

# LangChain 生态
langchain==1.2.11              # LangChain 框架
langchain-core==1.2.18         # LangChain 核心模块

# ============================================
# 数据处理
# ============================================
pandas==2.3.3                  # 数据分析与处理
numpy==2.2.6                   # 数值计算库
openpyxl==3.1.5                # Excel 文件读写

# ============================================
# 文档解析
# ============================================
PyPDF2==3.0.1                  # PDF 文档解析
python-docx==1.2.0             # Word 文档处理

# ============================================
# 向量检索与 RAG
# ============================================
faiss-cpu==1.13.2              # Facebook AI 相似性搜索（CPU版本）
sentence-transformers==5.2.3   # 句子嵌入模型

# ============================================
# 可视化
# ============================================
plotly==6.6.0                  # 交互式图表
matplotlib==3.10.8             # 静态图表
scipy==1.15.3                  # 科学计算（用于样条插值）

# ============================================
# GitHub 集成
# ============================================
PyGithub==2.8.1                # GitHub API 客户端

# ============================================
# 说明
# ============================================
# 安装命令: pip install -r requirements.txt
# 创建环境: conda create -n allenv python=3.10 -y
# ============================================
